Sprout a Lens module
See: https://github.com/zinc-collective/mp-plasticbullet/issues/4

As we learn how to Swift; we're stepping away from the SwiftUI/etc.
system and focusing more on building a library for doing the filtering.

We'll start with downloading a file from the Internet, saving it, then
running it through a LensStack, and saving a copy at each stage.

This way, we can start to diagnose if/how the filters are working
without futzing about with SwiftUI; which adds a confounding factor when
trying to figure things out.